+++ nano/core_test/peer_container.cpp
@@ -1,4 +1,4 @@
-#include <nano/node/socket.hpp>
+#include <nano/node/transport/socket.hpp>
 #include <nano/node/transport/tcp.hpp>
 #include <nano/node/transport/tcp_server.hpp>
 #include <nano/test_common/network.hpp>
