+++ nano/node/bootstrap/block_deserializer.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/blocks.hpp>
 #include <nano/node/bootstrap/block_deserializer.hpp>
-#include <nano/node/socket.hpp>
+#include <nano/node/transport/socket.hpp>
 #include <nano/secure/buffer.hpp>
 
 nano::bootstrap::block_deserializer::block_deserializer () :
