+++ nano/node/bootstrap/bootstrap_bulk_pull.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <nano/node/messages.hpp>
-#include <nano/node/socket.hpp>
+#include <nano/node/transport/socket.hpp>
 
 #include <unordered_set>
