+++ nano/node/bootstrap/bootstrap_connections.hpp
@@ -2,7 +2,7 @@
 
 #include <nano/node/bootstrap/bootstrap_bulk_pull.hpp>
 #include <nano/node/common.hpp>
-#include <nano/node/socket.hpp>
+#include <nano/node/transport/socket.hpp>
 
 #include <atomic>
