+++ nano/node/transport/message_deserializer.hpp
@@ -2,7 +2,7 @@
 
 #include <nano/node/common.hpp>
 #include <nano/node/messages.hpp>
-#include <nano/node/socket.hpp>
+#include <nano/node/transport/socket.hpp>
 
 #include <memory>
 #include <vector>
